Detect 5 GHz WLAN
Attempt to detect 5GHz WLAN so we can apply different RF
characteristics for it (shorter range). Note: Test phone
does not support 5GHz band so can not test. May be an issue
with upgrading on phones that have supported 5 GHz band and
have RF IDs (MAC addresses) saved with what is now a 2.4 GHz
type identification.

Signed-off-by: Tod Fitch <Tod@FitchDesign.com>